Create Number_guess_game
A number guessing game, in which the user selects a range. lets say User selected a range , i.e., from A to B, where A and B belong to integer. some random integer will be selected by the system and the user had to guess that integer in the minimum number of guesses. Here is problem statement , algorithm and source code to refer from. please refrain from copying the whole solution . We will suggest to only help and understand the flow of project.